Refactor FlatOpc to be extension methods
