+++ src/DocumentFormat.OpenXml.Framework/Features/IProgrammaticIdentifierFeature.cs
@@ -1,0 +1,9 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace DocumentFormat.OpenXml.Features;
+
+internal interface IProgrammaticIdentifierFeature
+{
+    string ProgramId { get; }
+}
